fix: Move format.rs
