+++ src/sstable/mod.rs
@@ -1,4 +1,3 @@
-pub mod format;
 mod index;
 pub mod manager;
 mod storage;
